fpay-flask/webapp/requirements.txt
Flask==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.4.5
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
numpy
pandas==0.25.1
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.6
